# src/main/java/com/ebb/journal/model/dto/TokenResponse.java
package com.ebb.journal.model.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.extern.jackson.Jacksonized;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Jacksonized
@Builder
public class TokenResponse {
  private String accessToken;
  private String idToken;
  private String refreshToken;
}